[WIP] Update Next.js SSR example
Backporting from reference arch
