+++ nextjs-ssr/checkout/package.json
@@ -3,13 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "node ./server.js",
     "build": "next build",
-    "start": "next start"
+    "start": "NODE_ENV=production node ./server.js"
   },
   "dependencies": {
     "nextjs-shared": "*",
-    "@module-federation/nextjs-ssr": "0.0.1-beta.21",
+    "@module-federation/nextjs-ssr": "0.0.3-beta.2",
     "lodash": "4.17.21",
     "next": "12.1.0",
     "next-compose-plugins": "^2.2.1",
