+++ nextjs-ssr/checkout/pages/index.js
@@ -1,3 +1,12 @@
-import { createFederatedCatchAll } from 'nextjs-shared';
+import dynamic from "next/dynamic";
+const page = import("home/home");
 
-export default createFederatedCatchAll(process.env.REMOTES);
+const Page = dynamic(() => import("home/home"));
+Page.getInitialProps = async (ctx) => {
+  const getInitialProps = (await page).default?.getInitialProps;
+  if (getInitialProps) {
+    return getInitialProps(ctx);
+  }
+  return {};
+};
+export default Page;
