+++ nextjs-ssr/home/async-pages/_app.js
@@ -1,0 +1,11 @@
+import Nav from '../components/nav';
+function MyApp({ Component, pageProps }) {
+  return (
+    <>
+      <Nav />
+      <Component {...pageProps} />
+    </>
+  );
+}
+
+export default MyApp;
