+++ nextjs-ssr/home/pages/_app.js
@@ -1,11 +1,12 @@
-import Nav from '../components/nav';
-function MyApp({ Component, pageProps }) {
-  return (
-    <>
-      <Nav />
-      <Component {...pageProps} />
-    </>
-  );
-}
+import dynamic from "next/dynamic";
+const page = import("../async-pages/_app");
 
-export default MyApp;
+const Page = dynamic(() => import("../async-pages/_app"));
+Page.getInitialProps = async (ctx) => {
+  const getInitialProps = (await page).default?.getInitialProps;
+  if (getInitialProps) {
+    return getInitialProps(ctx);
+  }
+  return {};
+};
+export default Page;
