+++ nextjs-ssr/home/server.js
@@ -1,0 +1,45 @@
+// server.js
+const { createServer } = require("http");
+const { parse } = require("url");
+const next = require("next");
+
+const dev = process.env.NODE_ENV !== "production";
+const hostname = "localhost";
+const port = process.env.PORT || 3000;
+// when using middleware `hostname` and `port` must be provided below
+const app = next({ dev, hostname, port });
+const handle = app.getRequestHandler();
+const cluster = require("cluster");
+
+const start = () =>
+  app.prepare().then(() => {
+    createServer(async (req, res) => {
+      try {
+        // Be sure to pass `true` as the second argument to `url.parse`.
+        // This tells it to parse the query portion of the URL.
+        const parsedUrl = parse(req.url, true);
+
+        await handle(req, res, parsedUrl);
+      } catch (err) {
+        console.error("Error occurred handling", req.url, err);
+        res.statusCode = 500;
+        res.end("internal server error");
+      }
+    }).listen(port, (err) => {
+      if (err) throw err;
+      console.log(`> Ready on http://${hostname}:${port}`);
+    });
+  });
+
+if (cluster.isMaster) {
+  // Fork workers.
+  for (let i = 0; i < 1; i++) {
+    cluster.fork();
+  }
+
+  cluster.on("exit", (worker, code, signal) => {
+    cluster.fork();
+  });
+} else {
+  start();
+}
