+++ nextjs-ssr/shop/async-pages/p/[...slug].js
@@ -1,0 +1,6 @@
+export default function PDP() {
+  return <h1>PDP!!!</h1>;
+}
+PDP.getInitialProps = async () => {
+  return {};
+};
